+++ HISTORY.rst
@@ -9,6 +9,7 @@ v0.28 (unreleased)
 * support ``__post_init_post_parse__`` on dataclasses, #567 by @sevaho
 * allow dumping dataclasses to JSON, #575 by @samuelcolvin and @DanielOberg
 * fix ``pydantic.compiled`` on ipython, #573 by @dmontagu and @samuelcolvin
+* add `StrictBool` type, #579 by @cazgp
 
 v0.27 (2019-05-30)
 ..................